+++ app/src/main/java/app/envelop/ui/login/LoginActivity.kt
@@ -59,6 +59,7 @@ class LoginActivity : BaseActivity() {
         messageManager.showError(
           when (it) {
             LoginViewModel.Error.LoginError -> R.string.login_error
+            LoginViewModel.Error.UsernameMissing -> R.string.login_username_error
           }
         )
       }