+++ app/src/main/java/app/envelop/ui/login/LoginViewModel.kt
@@ -1,6 +1,7 @@
 package app.envelop.ui.login
 
 import app.envelop.domain.LoginService
+import app.envelop.domain.LoginService.UnverifiedUsername
 import app.envelop.ui.BaseViewModel
 import app.envelop.ui.common.*
 import io.reactivex.rxkotlin.addTo
@@ -44,7 +45,7 @@ class LoginViewModel
           finishToMain.finish(Finish.Result.Ok)
         } else {
           Timber.w(it.throwable())
-          errors.onNext(Error.LoginError)
+          errors.onNext(if (it.throwable() is UnverifiedUsername) Error.UsernameMissing else Error.LoginError )
         }
       }
       .addTo(disposables)
@@ -62,7 +63,7 @@ class LoginViewModel
   fun errors() = errors.hide()!!
   fun finishToMain() = finishToMain.hide()!!
 
-  sealed class Error {
-    object LoginError : Error()
+  enum class Error {
+    LoginError, UsernameMissing
   }
 }