+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+- Error Message for Invalid Username
 - Socket Timeout Exceptions and GaiExceptions on unsubscribed RxSingles
 
 ## [1.2.8] - 2019-02-05